+++ aws-java-sdk-imagebuilder/src/main/java/com/amazonaws/services/imagebuilder/model/Distribution.java
@@ -36,7 +36,7 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
     private String region;
     /**
      * <p>
-     * The specific AMI settings (for example, launch permissions, AMI tags).
+     * The specific AMI settings; for example, launch permissions or AMI tags.
      * </p>
      */
     private AmiDistributionConfiguration amiDistributionConfiguration;
@@ -52,6 +52,12 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
      * </p>
      */
     private java.util.List<String> licenseConfigurationArns;
+    /**
+     * <p>
+     * A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * </p>
+     */
+    private java.util.List<LaunchTemplateConfiguration> launchTemplateConfigurations;
 
     /**
      * <p>
@@ -95,11 +101,11 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
 
     /**
      * <p>
-     * The specific AMI settings (for example, launch permissions, AMI tags).
+     * The specific AMI settings; for example, launch permissions or AMI tags.
      * </p>
      * 
      * @param amiDistributionConfiguration
-     *        The specific AMI settings (for example, launch permissions, AMI tags).
+     *        The specific AMI settings; for example, launch permissions or AMI tags.
      */
 
     public void setAmiDistributionConfiguration(AmiDistributionConfiguration amiDistributionConfiguration) {
@@ -108,10 +114,10 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
 
     /**
      * <p>
-     * The specific AMI settings (for example, launch permissions, AMI tags).
+     * The specific AMI settings; for example, launch permissions or AMI tags.
      * </p>
      * 
-     * @return The specific AMI settings (for example, launch permissions, AMI tags).
+     * @return The specific AMI settings; for example, launch permissions or AMI tags.
      */
 
     public AmiDistributionConfiguration getAmiDistributionConfiguration() {
@@ -120,11 +126,11 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
 
     /**
      * <p>
-     * The specific AMI settings (for example, launch permissions, AMI tags).
+     * The specific AMI settings; for example, launch permissions or AMI tags.
      * </p>
      * 
      * @param amiDistributionConfiguration
-     *        The specific AMI settings (for example, launch permissions, AMI tags).
+     *        The specific AMI settings; for example, launch permissions or AMI tags.
      * @return Returns a reference to this object so that method calls can be chained together.
      */
 
@@ -244,6 +250,76 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
     }
 
     /**
+     * <p>
+     * A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * </p>
+     * 
+     * @return A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     */
+
+    public java.util.List<LaunchTemplateConfiguration> getLaunchTemplateConfigurations() {
+        return launchTemplateConfigurations;
+    }
+
+    /**
+     * <p>
+     * A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * </p>
+     * 
+     * @param launchTemplateConfigurations
+     *        A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     */
+
+    public void setLaunchTemplateConfigurations(java.util.Collection<LaunchTemplateConfiguration> launchTemplateConfigurations) {
+        if (launchTemplateConfigurations == null) {
+            this.launchTemplateConfigurations = null;
+            return;
+        }
+
+        this.launchTemplateConfigurations = new java.util.ArrayList<LaunchTemplateConfiguration>(launchTemplateConfigurations);
+    }
+
+    /**
+     * <p>
+     * A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * </p>
+     * <p>
+     * <b>NOTE:</b> This method appends the values to the existing list (if any). Use
+     * {@link #setLaunchTemplateConfigurations(java.util.Collection)} or
+     * {@link #withLaunchTemplateConfigurations(java.util.Collection)} if you want to override the existing values.
+     * </p>
+     * 
+     * @param launchTemplateConfigurations
+     *        A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public Distribution withLaunchTemplateConfigurations(LaunchTemplateConfiguration... launchTemplateConfigurations) {
+        if (this.launchTemplateConfigurations == null) {
+            setLaunchTemplateConfigurations(new java.util.ArrayList<LaunchTemplateConfiguration>(launchTemplateConfigurations.length));
+        }
+        for (LaunchTemplateConfiguration ele : launchTemplateConfigurations) {
+            this.launchTemplateConfigurations.add(ele);
+        }
+        return this;
+    }
+
+    /**
+     * <p>
+     * A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * </p>
+     * 
+     * @param launchTemplateConfigurations
+     *        A group of launchTemplateConfiguration settings that apply to image distribution for specified accounts.
+     * @return Returns a reference to this object so that method calls can be chained together.
+     */
+
+    public Distribution withLaunchTemplateConfigurations(java.util.Collection<LaunchTemplateConfiguration> launchTemplateConfigurations) {
+        setLaunchTemplateConfigurations(launchTemplateConfigurations);
+        return this;
+    }
+
+    /**
      * Returns a string representation of this object. This is useful for testing and debugging. Sensitive data will be
      * redacted from this string using a placeholder value.
      *
@@ -262,7 +338,9 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
         if (getContainerDistributionConfiguration() != null)
             sb.append("ContainerDistributionConfiguration: ").append(getContainerDistributionConfiguration()).append(",");
         if (getLicenseConfigurationArns() != null)
-            sb.append("LicenseConfigurationArns: ").append(getLicenseConfigurationArns());
+            sb.append("LicenseConfigurationArns: ").append(getLicenseConfigurationArns()).append(",");
+        if (getLaunchTemplateConfigurations() != null)
+            sb.append("LaunchTemplateConfigurations: ").append(getLaunchTemplateConfigurations());
         sb.append("}");
         return sb.toString();
     }
@@ -294,6 +372,10 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
             return false;
         if (other.getLicenseConfigurationArns() != null && other.getLicenseConfigurationArns().equals(this.getLicenseConfigurationArns()) == false)
             return false;
+        if (other.getLaunchTemplateConfigurations() == null ^ this.getLaunchTemplateConfigurations() == null)
+            return false;
+        if (other.getLaunchTemplateConfigurations() != null && other.getLaunchTemplateConfigurations().equals(this.getLaunchTemplateConfigurations()) == false)
+            return false;
         return true;
     }
 
@@ -306,6 +388,7 @@ public class Distribution implements Serializable, Cloneable, StructuredPojo {
         hashCode = prime * hashCode + ((getAmiDistributionConfiguration() == null) ? 0 : getAmiDistributionConfiguration().hashCode());
         hashCode = prime * hashCode + ((getContainerDistributionConfiguration() == null) ? 0 : getContainerDistributionConfiguration().hashCode());
         hashCode = prime * hashCode + ((getLicenseConfigurationArns() == null) ? 0 : getLicenseConfigurationArns().hashCode());
+        hashCode = prime * hashCode + ((getLaunchTemplateConfigurations() == null) ? 0 : getLaunchTemplateConfigurations().hashCode());
         return hashCode;
     }